fix(tiled-png-writer): implement TODO entries 71-73
- Entry 71: replaced detached thread + manual promise with
  std::async(std::launch::async), matching the pattern in png-writer.cpp.
  Exceptions from writeToFile now propagate to the caller via the future.
  Both call sites in tab_maps.cpp already call .get(), so no callers need
  changing. Added doc comment noting caller must retain the future.

- Entry 72: updated tiles map comment to explicitly document insertion-order
  vs lexicographic difference and Porter-Duff blending order-dependence for
  semi-transparent tiles. No code change needed; WoW map tiles are opaque.

- Entry 73: changed Stats::expectedTiles from uint32_t to uint64_t.
  Updated getStats() to use static_cast<uint64_t>(tileCols) * tileRows,
  matching JS 64-bit Number semantics and eliminating uint32_t overflow.
  Removed unused <thread> include.

+++ src/js/tiled-png-writer.h
@@ -67,7 +67,7 @@ public:
 		uint32_t imageWidth;
 		uint32_t imageHeight;
 		uint32_t tileSize;
-		uint32_t expectedTiles;
+		uint64_t expectedTiles; // JS uses 64-bit Number; uint32_t overflows for large tile grids
 		double sparseRatio;
 	};
 
@@ -99,9 +99,9 @@ private:
 	uint32_t tileSize;
 	uint32_t tileCols;
 	uint32_t tileRows;
-	// JS equivalent uses Map() which preserves insertion order. std::map
-	// sorts by key, but the rendering is position-based (pixelX/pixelY) so
-	// the iteration order doesn't affect the final image. Using std::map
-	// for deterministic output.
+	// JS uses Map() (insertion-ordered); C++ uses std::map (lexicographic).
+	// For fully-opaque tiles this is irrelevant. For semi-transparent overlapping
+	// tiles, Porter-Duff blending is order-dependent and may differ from JS.
+	// WoW map tiles are typically fully opaque.
 	std::map<std::string, Tile> tiles;
 };
